Separated Wreck controller
